Update patch set 1
Patch Set 1:

(1 comment)

Thanks for the CL.

It was sent after the freeze and it's not fixing a bug or increasing stability, it is a code refactor. There doesn't seem to be any harm in leaving it to be looked at during Go 1.16 development cycle, so we can spend more time on CLs that are helping with stability. Adding wait-release hashtag.

Patch-set: 1
CC: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>