Update patch set 42
Patch Set 42:

(7 comments)

> Patch Set 41:
> 
> (1 comment)
> 
> Please address the test suite. Thanks.

@robertgriesemer

I just want to say that I've had eye strain for the whole week so I've struggled to be in front of a screen. I knew that you were super busy and this was low-priority so I tried my best to complete the task be the end of the week but it's gotten to the point where I can't seem spend even 15 minutes in front a screen.

It was an absolute honour working with you and learning from the best. I've done most of the tasks. The only remaining task is the hexadecimal test suite. Hopefully you'll be able to take it from here.

Regards,

-PJ

Patch-set: 42