runtime: do not attempt bulkBarrierPreWrite when dst slice length is zero
If dst slice length is zero then mallocgc is using a fast path to only return
a pointer to runtime.zerobase. There are no heapBits for that address readable
by bulkBarrierPreWriteSrcOnly which will cause a panic.

Protect against this by not calling bulkBarrierPreWriteSrcOnly if there is
nothing to copy. This is the case for all cases where the length of the
destination slice is zero.

runtime.growslice and runtime.typedslicecopy have fast paths that
do not call bulkBarrierPreWrite for zero copy lengths either.

Fixes #38929

Change-Id: I78ece600203a0a8d24de5b6c9eef56f605d44e99