net/http: remove redundant numeric StatusCode from Status message
In the case of parsing HTTP status with a message, eg: `HTTP/1.1 200 OK`,  `StatusCode` is correctly set to `200` but `Status` is parsed to `200 OK` when it should only be `OK`. This leads to inconsistency with `http.StatusText` which only returns `OK` instead of `200 OK`.

In the case `Status` includes a space, it includes a numeric status code we can trim out the numeral and move it to `StatusCode`

Fixes #38996

Change-Id: I64a64bbdf11ea5055a0a78a6ebc96bc200682e3b
GitHub-Last-Rev: a05be6cdf563ff4ee5b239b6352513c4a4696ef0
GitHub-Pull-Request: golang/go#38997