runtime: directly goready() in StopTrace
As a straw man fix for #39004, change StopTrace to goready() the reader
just like any other goroutine. This avoids the complication in schedule
to run the reader, and more importantly, the missed complication of
handling the reader on a locked m. Now, goready() and the scheduler will
be able to handle waking spinning Ms, etc just as with any other
goroutine.

This is presumably incorrect, or this is what would have been done all
along. Yet, I do not see why that is.

STW is excluding the ReadTrace critical section; a racing call to
ReadTrace that begins after the STW will never block and proceed to
shutdown as expected. goready() won't attempt to trace, since
trace.enabled has already been reset.

Please help me see where this doesn't work, or if it is now OK, let's
simplify things. :)

Fixes #39004

Change-Id: Ia0fa700ec2d0e88fd3bb3f42369b8ff80f99a710