container/list: check the list state before mutation
It is not enough to check if the Element's list pointer is equal to some
particular List pointer to answer whether an Element belongs to that List.
List might be zeroed and then mutated, which makes it completely different,
so that previous elements must not be connected with the new generation of
the list.

Fixes #39014

Change-Id: I8cd6753baeb86cbce0b0010d64c37275df0cd457
GitHub-Last-Rev: 04e5753c1ad8192ef6da2f24c85c5fc835d67b77
GitHub-Pull-Request: golang/go#39015