encoding/asn1: document what Unmarshal returns in rest
This change documents what Unmarshal returns in rest. Mainly this
conerns how trailing elements of SEQUENCES are treated. For context
Unmarshal treats trailing elements of a SEQUENCE that do not have
matching struct fields as valid, as this is how ASN.1 structures
are typically extended. This can be somewhat confusing as you might
expect those elements to be appended to rest, but rest is really only
for trailing data unrelated to the structure being parsed (i.e. if you
appened a second sequence to b, it would be returned in rest).

Fixes #35680 and #38545

Change-Id: Ia2c68b2f7d8674d09e859b4b7f9aff327da26fa0