net/http: handle RawPath in StripPrefix
The StripPrefix wrapper will strip a prefix string from the request's
URL.Path field, but it didn't touch the RawPath field. This leads to the
confusing situation where a request comes in with its URL.RawPath
populated (due to some escaped characters in the request path), and
RawPath contains the stripped prefix but Path does not.

This change modifies StripPrefix to strip the prefix from both Path and
RawPath. If there are escaped characters in the prefix part of the
request URL the stripped handler serves a 404 instead of invoking the
underlying handler with a mistmatched Path/RawPath pair.

This is a backward incompatible change for a very small minority of
requests; I would be surprised if anyone is depending on this behavior,
but it is possible. If that's the case, we could make a more
conservative change where the RawPath is trimmed if possible, but when
the prefix contains escaped characters then we don't 404 but rather send
through the invalid Path/RawPath pair as before.

Fixes #24366

Change-Id: I7030b8c183a3dfce307bc0272bba9a18df4cfe08