Update patch set 1
Patch Set 1:

Thanks for the change. The tree is currently closed for the Go 1.15 release (only CLs to stabilize that release are being merged), but we can go ahead and review this change with the intent to merge it when the tree reopens for Go 1.16.

I think this test would benefit from two things:


1. Please expand the change description (for GitHub PRs, the first comment on the PR) to describe the rationale for the change. Why do we believe that the coverage numbers are meaningful for failed tests?

I suspect that in practice the coverage for failed tests will be systematically lower than for passing tests, simply because many tests use `t.Fatalf` or similar on failure, which skips the remainder of the test (and any associated coverage for that test).

It might be preferable to file an issue (https://golang.org/issue/new) to discuss this in more detail. (Then you can refer to the issue by number in the change description to associate the change with the issue discussion.)


2. Please add an integration test (in src/cmd/go/testdata/script; see the README there for details) illustrating the expected benefit of this change.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>