Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi, I don't really understand how this change could improve runtime performance. Isn't this a thing which should be "fixed" in the compiler itself? Just curious as I was using aliased functions before.

Hi Leon,

Sorry for the late replying, just noticed the new comment.
Originally, 'block' is a function literal which is determined at runtime, not compile-time, so the compiler doesn't have sufficient information to determine if it's safe to allocate its parameter's (d) object on stack or not, thereby go with the conservative way to make it heap-allocated.

Patch-set: 4