cmd/compile: add test case for issue 38068
New test case for issue 38068, which deals with build reproducibility:
do a pair of builds, the first with the concurrent back end turned on,
and the second with -c=1, then check to make sure we get the same
output (using a test case that triggers late inlining into wrapper
methods).

Updates #38068.

Change-Id: I4afaf78898706a66985f09d18f6f6f29876c9017