Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (2 comments)
> 
> I don't see why we need these extra cases unless the code is incorrect for those arguments. If that's the case, they must be missing in the tests, and if they are missing then this code needs to add those tests.
> 
> For tanh, have you verified that for x == NaN the result is wrong?

I have verified that the code is correct for x == NaN. Whether it changes or not, the result is correct. I just think that this change seems to enhance code readability in special cases. Compared with other trigonometric and hyperbolic functions, they also deal with special cases in this way, whether it is better to keep the style consistent.

It doesn't matter if you think this is an unnecessary change. This is just a small suggestion and needs to be discussed.

Patch-set: 1