Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > (2 comments)
> > 
> > I don't see why we need these extra cases unless the code is incorrect for those arguments. If that's the case, they must be missing in the tests, and if they are missing then this code needs to add those tests.
> > 
> > For tanh, have you verified that for x == NaN the result is wrong?
> 
> I have verified that the code is correct for x == NaN. Whether it changes or not, the result is correct. I just think that this change seems to enhance code readability in special cases. Compared with other trigonometric and hyperbolic functions, they also deal with special cases in this way, whether it is better to keep the style consistent.
> 
> It doesn't matter if you think this is an unnecessary change. This is just a small suggestion and needs to be discussed.

Thanks for verifying.

The reason why the other transcendental functions handle special cases like this is because otherwise they return the wrong result.

In this case the result is correct without it, so adding an extra case makes the code slower: NaN is basically never an argument, so adding an extra test - while it may make it more obvious what happens - will slow then all other cases (which in reality are 100% of all cases).

Besides correctness, we want these routines to be as fast as possible. Adding an extra test just for readability is not justified in this case. Note also that actually fully understanding these routines requires significant numeric analysis, so adding this extra case for readability is not going to make a difference.

Please abandon this change. Thanks.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2