os: add a unit-test for creating a directory symlink before the directory exists
Based on failures seen on CL 234537, I suspected that there may be a
bug in our symlink handling on Windows. After a fair bit of debugging,
I discovered that the bug occurs when the symlink is created before
its target directory.

I did some more investigation, and learned that On Windows, the
CreateSymbolicLinkA system call requires a flag to indicate whether
the destination is a file or a directory: if
SYMBOLIC_LINK_FLAG_DIRECTORY is set, the symlink is a directory
symlink, and otherwise, it is a file symlink.

Our current implementation of os.Symlink on Windows uses a call to
os.Stat to determine which kind of link to create. Unfortunately, if
that Stat call fails, Symlink assumes that it is a file rather than
reporting the error to the caller. The result is that a sequence of
apparently-successful calls can create a symlink that appears to point
to the correct target but cannot actually be opened using os.Open.

This change adds a (failing) test illustrating the bug.

Change-Id: I04ed6440af008a17ca09410f5678b0a2e21257a3