go/types: replace Config.UsesCgo with Checker.CgoFiles
In CL 33677, I added a UsesCgo flag to indicate that the user will
provide us with _cgo_gotypes.go in files. This works, but it's a
little clumsy, because we can't be entirely sure which file it will
be. This isn't a problem under the current implementation, where we
just type-check everything altogether. But in the future, we might
want to separately type-check _cgo_gotypes.go, which would require
knowing which file that is, assuming it's even present.

This CL replaces Config.UsesCgo with a new Checker.CgoFiles entry
point, which allows the caller to separately provide the
_cgo_gotypes.go AST.

Config.UsesCgo was only added this development cycle, and only
recently, so it should be safe to change. Because we're in development
freeze, this CL is the minimum change to simply update the API, before
it's cemented in place by Go 1 compat.

Updates #16623.
Updates #39072.

Change-Id: Iaeb16fdc0dcbfa0d13b1a2b25bd6745e7f209441