Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Could we just delete the environment variable entirely, and just leave a constant to switch on/off the jump alignment?
> > 
> > My understanding is that we wanted to make it easy to back out, at least as far as reenabling the EV, in case we were in some way wrong about this.
> 
> Hmmm, my understanding was to make it easy to back out to not padding, instead of backing out to EV... Maybe we want to check with others?
> 
> Austin, what do you think? Thanks.

I think the problem with backing all the way out to no padding is that is a largish performance problem for that microcode patch.

But either way, someone with a definitive opinion needs to speak up, or this CL just waits and the knob stays in.

Patch-set: 2
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>