os: reject Symlink with a nonexistent oldname on Windows
The Windows CreateSymbolicLinkA system call requires a flag that
indicates whether the destination of the link is a file or a
directory. If the destination does not exist at all, we cannot set
that flag correctly. Instead, return an explicit error.

(Prior to this change, we would default to a file symlink when the
target did not exist, but if the destination turned out to actually be
a directory that would silently create a broken link.)

Also fix a bug in determining which path to stat when the target path
is relative to the volume (rather than the directory). Previously we
would erroneously prepend the link path, resulting in an invalid path
which was then masked by the lack of error-propagation.

Fixes #39183

Change-Id: Id1df2e1fc864462c03e7bfd30f84093fbc32ac1a