Update patch set 1
Patch Set 1:

I don't think this is an appropriate use of sync.Pool.  We don't know whether this makes sense for the application, and we shouldn't force the use of a Pool here.

A better approach would be to export the insert method, perhaps with appropriate restrictions.  Then the application could decide how to handle the memory usage of elements.

That said, I don't think we are going to be making any significant changes to the container/list package at this time.  For almost all uses, slices are better in practice.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>