Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I don't totally understand why we have these... If all variables of that type have been optimized out of a binary, what's the debugging going to do with the type? Are there useful comments or git history here that explains?

Hi Austin,

Prior to Go 1.9 when we generated DWARF for the params and local vars in a subprogram, the generation was done off of the post-optimization Dcl list for the function, meaning that if local variable "x" was completely optimized away, there would be no entry for it at all in the subprogram. So as a user if you did

(gdb) print x
No symbol "x" in current context.
(gdb)

which I think most folks agreed was not a good debugging experience. A better experience is to have the debugger "Variable 'x' value not available", or equivalent. 

When we added DWARF support for inlining, the problem was exacerbated (since local vars within inlined routine bodies tend to get deleted a lot). So as part of the inlining support we added this code:
  https://go.googlesource.com/go/+/0c0c3c186b1c1a7d3bc9d8a4aaec24767b5bfa81/src/cmd/compile/internal/gc/pgen.go#519

This led in short order to this bug:

  https://github.com/golang/go/issues/22941

which in turn led to this fix:

https://go.googlesource.com/go/+/9372166faa1bf04f0d83706da2b33ed2659ee976/src/cmd/compile/internal/gc/pgen.go#561

Note that this was back in the days when we still passed Autom's for each function all the way through to the linker.

Much later on when we wanted to get rid of Autom's, we moved to R_USETYPE as a replacement (I did that CL back this last September I recall).

Hope this helps.

Patch-set: 3
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>