Update patch set 3
Patch Set 3:

> Yeah, I think that was the context. Now I wonder, though, do we need to include the type for the dead variables? I think we still want the debugger to know there was a variable named x and is optimized out. But the type may not really matter.

It's true, it is kind of a grey area.  You can argue that if the types aren't included you might somehow lose the last reference to a type, and then not be able to use that type in a debugging session... but this is not a super strong argument I think. 

I suppose I cam get rid of R_USETYPE creation entirely and see what breaks?

Patch-set: 3
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>