Update patch set 7
Patch Set 7:

This isn’t a backwards compatible change. Walk is documented as “All errors that arise visiting files and directories are filtered by walkFn.“ but this breaks that so that errors that occur reading directory names are no longer filter by walkFn. (I believe that’s even a common occurrence for directories with restrictive permissions).

Patch-set: 7
CC: Gerrit User 20190 <20190@62eb7196-b449-3ce5-99f1-c037f21e1705>