crypto/x509: enforce algorithmIdentifier matching/encoding restrictions
When parsing a certificate enforce that the certificate
signatureAlgorithm field matches the tbsCertificate signature field,
as specified in RFC 5280, and enforce the contents of
algorithmIdentifier parameter fields for supported signature algorithms.
If a algorithmIdentifier contains unexpected content in the parameter
field ParseCertificate returns UnknownSignatureAlgorithm. At some point
it might make sense to fail out of parsing in this case, since the
certificate cannot be used with an unknown signature algorithm (without
the user messing around with its contents) but that is outside the scope
of this change.

Fixes #21118

Change-Id: I143df050c38241691ea3bdd6506a56e3bf526030