crypto/x509: enforce algorithmIdentifier matching/encoding restrictions
When parsing a certificate enforce that the two algorithm identifier
fields in the outer certificate and the inner tbsCertificate match,
and that the parameters fields of these algorithm identifiers are
properly encoded. Also reject certificates that use unknown signature
algorithms, rather than returning UnknownSignatureAlgorithm in the
SignatureAlgorithm field, as they cannot be used for anything.

Fixes #21118

Change-Id: I143df050c38241691ea3bdd6506a56e3bf526030