Update patch set 4
Patch Set 4:

Could y'all please advise how we should proceed with this CL?
On the issue tracker it is tagged for Go1.15, but the CL was sent past
the deadline. However, it is a security issue for which other versions
have no mitigation. Also there is going to be some work for HTTP/2.

Thank you.

Patch-set: 4
CC: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 28759 <28759@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>