Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> This isn’t a backwards compatible change. Walk is documented as “All errors that arise visiting files and directories are filtered by walkFn.“ but this breaks that so that errors that occur reading directory names are no longer filter by walkFn. (I believe that’s even a common occurrence for directories with restrictive permissions).

But it should handle the SkipDir, there is a same issue, https://github.com/golang/go/issues/16280

So it changed as


err1 := walkFn(path, info, nil)
if err1 != nil {
   if err1 == filepath.SkipDir{
      return err1
   }
}
names, err := readDirNames(path)
if err != nil {
}

it will be a backwards compatible change.

The problem is it will be a large amount of memory if the directory has a large number of files.
The skipdir would not walk.

Patch-set: 7