Update patch set 3
Patch Set 3:

> > I suppose I cam get rid of R_USETYPE creation entirely and see what breaks?

I did an experiment in which I linked a small program with R_USETYPE relocations completely turned off (this is not quite a himom.go, but close). It looks as though there are a fair number of types that wind up being omitted, on the order of ~200. I can provide the list if anyone is interested.

In addition, all.bash doesn't pass -- in runtime-gdb_test.go there is a testpoint added for https://github.com/golang/go/issues/17830 that is failing:

    runtime-gdb_test.go:480: could not find []main.astruct; in 'info typrs astruct' output

Looking at the issue in question, it seems that from the perspective of the Delve folks we still need to have R_USETYPE or something like it.

Patch-set: 3