Update patch set 1
Patch Set 1:

A bit of a note: the 5280 sig alg check is a bit best effort because we don't have easy access to the actual bytes to do a byte-for-bytes equality check. Because of how encoding/asn1 treats sequences with trailing fields either the tbsCertificate or certificate algorithmIdentifier SEQUENCEs could contain a bunch of extra junk, and we wouldn't catch it here.

Patch-set: 1