net/mail: fix ParseDate obsolete timezone "T" index check
Fixes the check for RFC 5322 "obsolete time zone" to ensure
that we correctly extract the entire date from the "T" of
the implied time zone.

Obsolete Time zones come in the form:
* GMT
* PST
* MDT
etc, as per Section 4.3 of RFC 5322,
https://tools.ietf.org/html/rfc5322#section-4.3.

The prior check from CL 117596 erronenously used strings.Index
which selects the first "T", and that meant that dates containing
days "Tue" or "Thu" could not be parsed.

This change uses strings.LastIndex to correctly retrieve the whole date string.

Fixes #39260

Change-Id: I6d59d99bc4f05a82582c826b5c5a080a25fd999b