cmd/compile: when register map is not used, treat it as invalid
When go115ReduceLiveness is true (so we don't emit actual
register maps), RegMapValid can still be true as we use
regMapIndex to store unsafe point information. This causes the
compiler to crash when doing -live=2 debug print.

This CL let RegMapValid be false if go115ReduceLiveness is true,
as it is not an actual register map index. It is safe as
RegMapValid is only used in debug print.

Fixes #39251.

Change-Id: Ice087af491fa69c413f8ee59f923b72d592c0643