os: document Symlink idiosyncracies on Windows
Also propagate any Stat errors from Symlink if they are not equivalent
to either os.ErrNotExist os windows.ERROR_INVALID_NAME. I could not
figure out how to provoke such an error, but I could envision one
occurring if (say) a network drive is disconnected or lacks
credentials, and in that case an explicit failure seems preferable to
a permanently-broken symlink.

This is an alternative to CL 234879, although personally I prefer the
approach in CL 234879 over this one.

No test because I could not figure out how to provoke a non-excluded
error without setting up a network share, and setting up a network
share on the builders seems too invasive.

Fixes #39183

Change-Id: I5f31001fb9936cb2f340187a5a44048bb9ac45ba