Update patch set 7
Patch Set 7:

I've sent a new patch-set; there was a merge conflict due to some other test being added and I just rebased with master.

Also, do we have any feedback on how we'd like to proceed with this CL? I know it's past the deadline, but the issue has been assigned to the Go 1.15 milestone, and labelled "Security".

Thank you for your time!

Patch-set: 7