net/http: add examples for disabling HTTP/2 in Transport
The documentation that recommends disabling HTTP/2 in a
Transport just by setting TLSNextProto to a blank but
non-nil map doesn't directly apply to a clone from
http.DefaultTransport, because .Clone() only guarantees
exporting external fields. When the DefaultTransport
is created and used, there are fields that will be
set that enable HTTP/2, if GODEBUG=http2client=1+ is not set.

This change provides examples that show how to enable or disable
HTTP/2 on a Transport.

Fixes #39302

Change-Id: Ie8be1ff0e15fb64c44694a010758fb13c47d3ad8