Update patch set 3
Patch Set 3:

> I assumed you'd submit it when the gates opened up.

Agh, I didn't notice the change in the issue's milestone. I was just looking at the contents of this CL, and didn't see a R=1.16 or wait-release hashtag, which I assumed were the ways to mark a CL as not being OK to merge during the freeze.

I would have probably not merged this later in the freeze, but given that we're still in the first month, I thought it would be well within reason. It's a pretty safe bugfix, in my opinion, and I highly doubt any programs would break. If that happens, we still have the betas and RCs as safeguards.

Ian or Andy, any thoughts? Happy to be overruled and reverted if you think my judgement is off :)

Patch-set: 3
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 22285 <22285@62eb7196-b449-3ce5-99f1-c037f21e1705>