cmd/go: assign missing import errors to importing package
MissingImportErrors are created on the imported package, but they
should instead be reported in the importing package. Those errors
belong bettor on the importing package because it is the importer's
"fault" that the package that they are importing doesn't exist,
either because the import path doesn't exist at all, or because it
doesn't exist in the module being imported.

The error is created on the dummy imported package because it's determined
that that package doesn't exist while the imported package is being
built. To move the error to the correct package, we have it implement
a method "ParentError" that returns an error that can be assigned
to the parent. Then when the imports are being checked in internal/load.load,
errors matching that interface can be bubbled up.

One alternative to creating this interface could be to move the definition
of ImportMissingError to internal/load, so that internal/load could
use it without creating a package graph cycle between internal/load and
internal/modload.

Fixes #26909

Change-Id: Id97451c03af193df9daffe2f7520e47bbd256c12