Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > I assumed you'd submit it when the gates opened up.
> 
> Agh, I didn't notice the change in the issue's milestone. I was just looking at the contents of this CL, and didn't see a R=1.16 or wait-release hashtag, which I assumed were the ways to mark a CL as not being OK to merge during the freeze.
> 
> I would have probably not merged this later in the freeze, but given that we're still in the first month, I thought it would be well within reason. It's a pretty safe bugfix, in my opinion, and I highly doubt any programs would break. If that happens, we still have the betas and RCs as safeguards.
> 
> Ian or Andy, any thoughts? Happy to be overruled and reverted if you think my judgement is off :)

For the Go core team, we’ve been filing issues with a “freeze exception” prefix to get it looked at. https://golang.org/issues/38889 is an example of this. The underlying motivation is to ensure that any decisions we made to land things after the freeze were discussed openly. We haven’t formalized this at all, though, which is why you likely didn’t know about it :)

Dmitri is running our internal test suites on all our changes, so we’ll see if anything pops up as a result of this. We can always revert it if a problem pops up. For now I’d say it’s okay as it’s a bugfix and not feature work.

Patch-set: 3
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>